Update testset & pretrain model
